c est quasi bon

+++ Cercle_3/fdf/srcs/convertor.c
@@ -6,7 +6,7 @@
 /*   By: bsavinel <bsavinel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/17 15:53:14 by bsavinel          #+#    #+#             */
-/*   Updated: 2022/01/24 17:47:06 by bsavinel         ###   ########.fr       */
+/*   Updated: 2022/01/25 16:32:42 by bsavinel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,10 +39,11 @@ int	malloc_map2d(t_map *map)
 t_2Dpoint	isometric_point(t_3Dpoint p3d, int zoom, int scaling)
 {
 	t_2Dpoint	p2d;
+	//float tmpz;
 
 	p3d.x *= zoom;
 	p3d.y *= zoom;
-	p3d.z *= zoom / scaling;
+	p3d.z *= (float) zoom / scaling;
 	p2d.x = (p3d.x - p3d.y) * cos(0.463646716);
 	p2d.y = (p3d.x + p3d.y) * sin(0.463646716) - p3d.z;
 	return (p2d);
@@ -59,7 +60,8 @@ int	convertor(t_map *map)
 		i = 0;
 		while (i < map->len_line)
 		{
-			map->map2d[j][i] = isometric_point(map->map3d[j][i], map->zoom, map->scaling);
+			map->map2d[j][i] = isometric_point(map->map3d[j][i], map->zoom,
+					map->scaling);
 			i++;
 		}
 		j++;
